+++ server/src/main/resources/templates/error/404.html
@@ -1,0 +1,8 @@
+<!DOCTYPE html>
+<html lang="en">
+<body>
+<h1>Colloc 404 | Something went wrong! </h1>
+<h2>Our Engineers are on it</h2>
+<a href="/">Go Home</a>
+</body>
+</html>
