+++ server/src/test/kotlin/com/mobilejazz/colloc/CollocInteractorTest.kt
@@ -1,0 +1,46 @@
+package com.mobilejazz.colloc
+
+import com.mobilejazz.colloc.domain.interactor.CollocClassicInteractor
+import com.mobilejazz.colloc.domain.interactor.CollocInteractor
+import com.mobilejazz.colloc.domain.interactor.DownloadFileInteractor
+import com.mobilejazz.colloc.domain.model.Platform
+import kotlinx.coroutines.runBlocking
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.assertThrows
+import java.io.File
+
+class CollocInteractorTest {
+    private fun getInteractor(): CollocInteractor {
+        return CollocInteractor(
+            DownloadFileInteractor(),
+            CollocClassicInteractor(),
+        )
+    }
+
+    @Test
+    fun `no id returns null`() {
+        runBlocking {
+            assertThrows<CollocInteractor.Error.InvalidIdException> {
+                getInteractor()("", listOf(Platform.IOS))
+            }
+        }
+    }
+
+    @Test
+    fun `correct id generates a file`() {
+        runBlocking {
+            val id = "1FYWbBhV_dtlSVOTrhdO2Bd6e6gMhZ5_1iklL-QrkM2o"
+            val result = getInteractor()(id, listOf(Platform.ANGULAR))
+            assert(result is File)
+        }
+    }
+
+    @Test
+    fun `no platforms returns an error`() {
+        runBlocking {
+            assertThrows<CollocInteractor.Error.InvalidPlatformException> {
+                getInteractor()("some random string", listOf())
+            }
+        }
+    }
+}
